Task2/requirements.txt
torch>=2.0.0
transformers>=4.25.0
datasets>=2.8.0
evaluate>=0.4.0
seqeval>=1.2.0
tqdm>=4.64.0
numpy>=1.21.0
pillow>=9.0.0
matplotlib>=3.5.0